amsvh/requirements.txt
asgiref==3.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Markdown==3.4.1
mysql-connector==2.2.9
mysqlclient==2.1.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.5.0
PyMySQL==1.0.3
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
urllib3==1.26.12
yapf==0.32.0
